New Components:
* CellValue
What is the value of the cell?
* CellPos
Where on the grid is the cell?
* CellType
Is a cell one that is in the starting configuration? Is it free to change? Was it locked by the user?
* State
Not sure if this one will be used yet, but minor global state

Component Changes:

* Sprite
Takes an attribute to set while drawing the sprite (bold, underline, etc)

New Systems:

* SudokuLogic
Contains logic to handle inputs to the Sudoku grid. It will modify the cells as necessary (ex: highlight currently selected cell)